# api-definitions/Repository.openapi+json.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Repository API",
        "version": "v1"
    },
    "paths": {
...
        "/repository/players/{playerId}": {
            "get": {
                "tags": [
                    "Players"
                ],
                "parameters": [
                    {
                        "name": "playerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Players"
                ],
                "parameters": [
                    {
                        "name": "playerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}